refactor: use network message consistently [SF-969]
