+++ src/components/atoms/SupportedNetworks/index.ts
@@ -1,0 +1,1 @@
+export * from './SupportedNetworks';
