+++ src/components/organisms/Header/Header.tsx
@@ -3,16 +3,21 @@ import { useRouter } from 'next/router';
 import { useDispatch, useSelector } from 'react-redux';
 import SFLogo from 'src/assets/img/logo.svg';
 import SFLogoSmall from 'src/assets/img/small-logo.svg';
-import { Button, HighlightChip, NavTab } from 'src/components/atoms';
+import {
+    Button,
+    HighlightChip,
+    NavTab,
+    SupportedNetworks,
+} from 'src/components/atoms';
 import { HamburgerMenu, MenuPopover } from 'src/components/molecules';
 import { WalletDialog, WalletPopover } from 'src/components/organisms';
 import useSF from 'src/hooks/useSecuredFinance';
 import { setWalletDialogOpen } from 'src/store/interactions';
 import { RootState } from 'src/store/types';
-import { getEnvShort, getSupportedChainIds } from 'src/utils';
+import { getEnvShort } from 'src/utils';
 import { AddressUtils } from 'src/utils/address';
 import { isChipVisibleForEnv, isProdEnv } from 'src/utils/displayUtils';
-import { mainnet, sepolia, useAccount } from 'wagmi';
+import { mainnet, useAccount } from 'wagmi';
 
 const PRODUCTION_LINKS = [
     {
@@ -49,12 +54,6 @@ const HeaderMessage = ({
     chainId: number;
     chainError: boolean;
 }) => {
-    const chainIds = getSupportedChainIds();
-    const networkNames = [sepolia, mainnet]
-        .filter(chain => chainIds.includes(chain.id))
-        .map(chain => chain.name)
-        .map(name => (name === 'Ethereum' ? 'Mainnet' : name));
-
     if (chainId) {
         if (chainError) {
             return (
@@ -62,13 +61,7 @@ const HeaderMessage = ({
                     className='typography-caption-2 w-full bg-red p-[1px] text-center text-neutral-8'
                     data-testid='testnet-alert'
                 >
-                    Secured Finance only supported on{' '}
-                    <span className='capitalize'>
-                        {networkNames.length === 2
-                            ? networkNames.join(' and ')
-                            : networkNames.join(', ')}
-                    </span>{' '}
-                    in Ethereum
+                    <SupportedNetworks />
                 </div>
             );
         } else if (chainId !== mainnet.id) {
